move widget folder to utils local
